+++ src/normcap/__init__.py
@@ -1,3 +1,3 @@
 """Normcap Package."""
 
-__version__ = "0.2.8"
+__version__ = "0.2.9"